Separate window from logical_range_filter into an independent file
